# Container performance improvements