--- a/agent/main.go
+++ b/agent/main.go
@@ -143,6 +143,7 @@ func main() {
 
 func initMetadataProviders(cfg *config.AgentConfig) {
 	if err := docker.InitDockerUtil(&docker.Config{
+		CacheDuration:  cfg.ContainerCacheDuration,
 		CollectHealth:  cfg.CollectDockerHealth,
 		CollectNetwork: cfg.CollectDockerNetwork,
 		Whitelist:      cfg.ContainerWhitelist,