--- a/checks/container_test.go
+++ b/checks/container_test.go
@@ -72,7 +72,11 @@ func TestContainerChunking(t *testing.T) {
 }
 
 func BenchmarkAllContainers(b *testing.B) {
-	docker.InitDockerUtil(&docker.Config{CollectHealth: true, CollectNetwork: true})
+	docker.InitDockerUtil(&docker.Config{
+		CacheDuration:  10 * time.Second,
+		CollectHealth:  true,
+		CollectNetwork: true,
+	})
 	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {