Improves the performance of the container checks significantly:

(from a test w/ 70 running containers, 10s cache)

```
benchmark                    old ns/op     new ns/op     delta
BenchmarkAllContainers-2     78556378      3733144       -95.25%

benchmark                    old allocs     new allocs     delta
BenchmarkAllContainers-2     38768          6330           -83.67%

benchmark                    old bytes     new bytes     delta
BenchmarkAllContainers-2     7538615       1143256       -84.83%
```

Changes included:

* Removing stats we don't use from CPU and Memory to cut down on file reads. Caching MemLimit and CPULimit after the first run. (https://github.com/DataDog/datadog-process-agent/commit/dca5876e58e62ab5260104ef7aafc14c55fc9a77)
* Adding caching to container list metadata so we only get the latest containers every 10s. (https://github.com/DataDog/datadog-process-agent/commit/0505470c23a5c8e757f94a585464767d0ceddfc3)
* General alloc optimizations in file reading, etc. (https://github.com/DataDog/datadog-process-agent/commit/e44b0de7b9eb7f05c96a2dbf91e763e931bf6174)
* Parsing the docker health from the container list API. This allows us to remove the CollectContainerHealth option and have it on by default, because there is no perf impact. (https://github.com/DataDog/datadog-process-agent/commit/596537b38c05047d28f8cfc3c805954486714a9f)